requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.7.2
shap==0.45.1
joblib==1.3.2
catboost==1.2.2
aiogram==3.4.1
httpx==0.27.0
python-dotenv==1.0.1
pytest==8.0.2
